+++ src/bda/plone/payment/debit_order/__init__.py
@@ -10,7 +10,7 @@ _ = MessageFactory('bda.plone.payment')
 
 class DebitOrder(Payment):
     pid = 'debit_order'
-    label = _('debit_order', 'Debit Order')
+    label = _('debit_order', default=u'Debit Order')
 
     def init_url(self, uid):
         return '%s/@@debit_order?uid=%s' % (self.context.absolute_url(), uid)